Switch to pytest
pytest makes it easier to write, read, run and debug tests.  When using
pytest, there is no need to write custom test runner, and pytest has
builtin colored output.

This patch is removing the custom test runner defined in setup.py, and
use py.test for running the tests. pytest support unittest.TestCase
based tests so we don't have to do change the current tests.

This patch introduces a new build dependency, but since we are using
tox, this dependency is transparent to developers.

Change-Id: Idc078ed6f7efad2ba380e0de8cfae0d745f0aafd
Signed-off-by: Nir Soffer <nsoffer@redhat.com>