Simplify the tests using pytest facilities
When using pytest, there is no need to subclass unittest.TestCase or use
custom assert* methods. Instead, we can use plain assert.

For example, instead of writing:

    self.assertIn("sys", sys.modules)

We can write:

    assert "sys" in sys.modules

Unlike unittest and nose, when plain assert fails, pytest display the
failing code, the expected and actual output, and the difference between
them.  pytest also display a traceback of all the frames including
locals values, which make it much easier to debug failing tests.

pytest is using the test class names to detect test case classes. The
default pattern is Test*, so the classes were renamed to match the
default, which makes the code more familiar to new developers.

Since we don't use custom assert* methods, this change also fix the
compatibility with Python 2.6, since the tests used assertIn() which is
available only on Python 2.7.

Change-Id: I0dcdb3d0b5137009ceffeb0ce35769fc43921c91
Signed-off-by: Nir Soffer <nsoffer@redhat.com>